+++ weave-js/src/core/ops/domain/util.ts
@@ -1,5 +1,3 @@
-import {dict, list, typedDict, union} from '../../model';
-
 /**
  * `connectionToNodes` is a helper function that converts a `connection` type
  * returned from gql to its list of nodes. In many, many location in our
@@ -43,27 +41,3 @@ export const connectionToNodes = <T>(connection: MaybeConnection<T>): T[] =>
   (connection?.edges ?? [])
     .map(edge => edge?.node)
     .filter(node => node != null) as T[];
-
-const traceFilterPropertyTypes = {
-  trace_roots_only: union(['none', 'boolean']),
-  op_names: union(['none', list('string')]),
-  input_refs: union(['none', list('string')]),
-  output_refs: union(['none', list('string')]),
-  parent_ids: union(['none', list('string')]),
-  trace_ids: union(['none', list('string')]),
-  call_ids: union(['none', list('string')]),
-  wb_user_ids: union(['none', list('string')]),
-  wb_run_ids: union(['none', list('string')]),
-};
-
-export const traceFilterType = union([
-  'none',
-  typedDict(traceFilterPropertyTypes, Object.keys(traceFilterPropertyTypes)),
-]);
-export const traceLimitType = union(['none', 'number']);
-export const traceOffsetType = union(['none', 'number']);
-export const traceSortByType = union([
-  'none',
-  list(typedDict({field: 'string', direction: 'string'})),
-]);
-export const traceQueryType = union(['none', dict('any')]);
